+++ scripts/export/export.py
@@ -54,7 +54,7 @@ def get_project_builder() -> ProjectBuilder:
 def main():
     if not InstallationChecker.check():
         sys.exit(1)
-    if not update_i18n_files(clean=True, generate_doc=False):
+    if not update_i18n_files(clean=False, generate_doc=False):
         logger.error('You must update the translations.')
         sys.exit(1)
     if not get_project_builder().run():